Substantial progress towards implementation of clamping/associated boundary condition semantics.
Thought that simply defining reference points on an Assembly would be sufficient
to define surface boundary conditions. This is obviously not the case. Realized that it
is necessary to partition a surface, so a new face is generated on the surface, in
order to specify boundary conditions. Basically, it's necessary for a Face object to
exist in the desired region in order to specify a boundary condition on the region.