Commitment almost working.
Working on bug related to Abaqus' (poor) handling of open ODBs. The bug
signature is, sometimes, upon calling materialFromODB(), Abaqus will
give up and say that the "ODB is out of date and needs to be updated"
even though an absolute path to the ODB is passed to the
materialFromODB() function. My hypothesis is that Abaqus is not doing a
good job closing ODB objects when things like materialFromODB(),
partFromODB(), etc. are called.