More commitment progress.
There appears to be a lingering timing-related bug. There are a number
of bug signatures, but I think they all boil down to a situation where
an ODB is still being written to and something tries to import or create
something based on the content of the ODB. The bug signatures are at
least the following: segfault within Abaqus code, warning which looks
like "ODB is out of date, it's necessary to update it" when calling
something like materialFromOdb().